+++ alacritty_terminal/src/event_loop.rs
@@ -342,9 +342,14 @@ where
 
                         token if token == self.pty.child_event_token() => {
                             if let Some(tty::ChildEvent::Exited) = self.pty.next_child_event() {
-                                if !self.hold {
+                                if self.hold {
+                                    // With hold enabled, make sure the PTY is drained.
+                                    let _ = self.pty_read(&mut state, &mut buf, pipe.as_mut());
+                                } else {
+                                    // Without hold, shutdown the terminal.
                                     self.terminal.lock().exit();
                                 }
+
                                 self.event_proxy.send_event(Event::Wakeup);
                                 break 'event_loop;
                             }